Limit joint limits (solves planning problems)
This commit restricts the joints to the actually reachable angles.
Without this commit, planning between the following joint angles times
out with a "plan not found" error, even though there is a valid IK
solution for them:

   joint_group_positions{-2.811, 0.336, -0.666, -1.459, 1.306, 1.778};
   joint_group_positions = {-2.938, 3.007, -0.999, -0.685, -1.334, -1.242};

The reason is that it's impossible for the ur5_shoulder_lift_joint to go
from 0.336 to 3.007 without colliding with the robot structure, even though
both joint angles are technically within the joint angles of the robot.
Similarly, the ur5_elbow_joint can never reach values outside of [-pi,
pi] without self-collision. This makes the planning space discontinuous.

After this commit, the planning for the joint targets above now fails
immediately instead of timing out, because the joint angles are outside
the joint limits. Also, IK will now never return an unreachable set of
joint angles.