added improved spawn_powerdrill.launch
which uses a single argument to specify that the object should be
spawned relative to another object (e.g. a table), and that the
given coordinates are relative to it.

Not completely compatible with the old version (it's not possible
to wait for a table but give the coordinates in the world frame),
hence the new file.