introduced prefix parameters to urdf, srdf and config files
This allows us to compile a new robot description for every instance
of a robot we want to deploy using xacro, which will include a prefix
for every joint and link. Therefore many config files need to be
parameterized, too. All in all this makes it possible to set a tf_prefix
for our robot without using the inconsistently used and therefore not
correctly working rosparam 'tf_prefix'.