Fix astra + pico flexx frame_id with gazebo_plugins 2.9.2
In gazebo_plugins 2.9.2, the behavior of the gazebo_ros_openni_kinect
plugin was changed in this commit:

https://github.com/ros-simulation/gazebo_ros_pkgs/commit/34b2c27936cf1b00801b774935afb7a8e17e9c8c

It now no longer prepends the robotNamespace if the parameter tf_prefix
is empty. Since tf_prefix has been deprecated for a long time, we always
leave it empty. This means we have to prepend the prefix ourselves.

NOTE: This makes planning slower again, because the sensor data is now
properly integrated into the obstacle map again. Planning was only
faster temporarily because we didn't process any obstacles from the
environment.