moveit cfg: Add shape_transform_cache_lookup_wait_time
There are repeated error messages from moveit like this:

[ERROR] [1622559146.727209001, 248.826000000]: Transform error: Lookup would require extrapolation -1.783000000s into the future.  Requested time 248.600000000 but the latest data is at time 246.817000000, when looking up transform from frame [mobipick/base_link] to frame [mobipick/gripper_astra_depth_optical_frame]
[ERROR] [1622559146.727963287, 248.826000000]: Transform cache was not updated. Self-filtering may fail. If transforms were not available yet, consider setting robot_description_planning/shape_transform_cache_lookup_wait_time to wait longer for transforms

This commit adds the parameter shape_transform_cache_lookup_wait_time to
the config, but leaves it at the default value of 0.05. This allows
tuning the parameter later, although the reported time differences are
too large to fix with this param.