pick_n_place: Fix table height
The table height is used during placing. The calculation before was
incorrect (table height = 0.385 instead of 0.73). This did not matter on
the real robot (because the z pose of the table is set to 0.73/2, so the
total height was 0.73), but it broke placing in Gazebo and in the MoveIt
demo mode.