Remove roll rotation from powerdrill_with_grip
The roll=pi rotation in the URDF mean that the Gazebo pose and the mesh
pose were different. The object poses estimated by DOPE are based on the
mesh pose without any additional rotation. So in order to return the
same poses as DOPE on the real robot, the gazebo_object_publisher had to
apply another roll=pi rotation to the Gazebo pose. The logical_camera
plugin does not do this, so it returned incorrect poses.

This commit solves that problem by removing the roll=pi rotation from
the URDF, thereby making any special handling unnecessary.