moveit: Use depthimage instead of pointcloud in gazebo
On the real robot, we use the depthimage_octomap_updater. On Gazebo, we
used to use the pointcloud_octomap_updater instead, because in earlier
versions of Gazebo (Kinetic?), the simulated depth images had a wrong
time stamp, so the octomap updates would be shifted when the camera was
moving.

I've verified that this does not happen any more using the
filtered_cloud debug topic, so now we can switch to the
depthimage_octomap_updater. This has the following advantages:

* makes the software we run on Gazebo more similar to the real robot
* depthimage_octomap_updater is faster
* doesn't produce annoying TF errors (see DFKI-NI/mobipick_tables_demo#5)